requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.51.0
gunicorn==22.0.0
html5lib==1.1
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
PyJWT==2.8.0
PyPDF2==3.0.1
pyphen==0.15.0
pytest==8.2.0
pytest-django==4.8.0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
weasyprint==62.1
webencodings==0.5.1
zopfli==0.2.3
